trading_bot/requirements.txt
fastapi
uvicorn
python-binance
typer
streamlit
python-dotenv
pydantic
requests
